+++ config/routes.rb
@@ -16,6 +16,7 @@ Rails.application.routes.draw do
   end
   resources :rounds, only: [:show, :edit, :update] do
     resources :screenings, only: [:create]
+    resource :movie, only: [:edit, :update], module: 'rounds'
   end
 
   resources :screenings, only: [:destroy]
